Final adjustments for the demo

+++ pyoctowight/octo.py
@@ -67,11 +67,13 @@ def knockout_round():
 
     json_next = jsonify_prolog(next_round)
     winners = json.loads(json_next)
-    r = {'next_url': url_for('knockout_round')}
-    if isinstance(winners, basestring):
-        r['winner'] = winners
+    if isinstance(winners, basestring) or len(winners) == 0:
+        r = {'winner':  winners,
+             'next_round': url_for('winner', winner=winners)}
     else:
-        r['winners'] = winners
+        r = {'winners': winners,
+             'next_url': url_for('knockout_round')}
+
     json_response = json.dumps(r)
     return json_response
 
@@ -82,37 +84,41 @@ def premier_index():
 
 @app.route('/premier_reset')
 def premier_reset():
-    os.unlink(PREMIER_DATAFILE)
+    if os.path.isfile(PREMIER_DATAFILE):
+        os.unlink(PREMIER_DATAFILE)
     return ''
 
 @app.route('/premier', methods=['POST'])
 def premier():
     leagues = request.json['leagues']
     matches = request.json['matches']
-    all_leagues = [leagues[group] for group in sorted(leagues)]
+    league = [leagues[group] for group in sorted(leagues)][0]
     matches_played = [
         "match_played({}, {}, {}, {}, {}).".format(match['team1'].lower(), match['team2'].lower(),
                                                match['score1'], match['score2'],
                                                    unique_id())
         for match in matches]
-    data = """leagues({}).
+    data = """league({}).
     {}
-    """.format(prologify_json(json.dumps(all_leagues)),
+    """.format(prologify_json(json.dumps(league)),
                "\n".join(matches_played))
 
-    winners = call_prolog('premier.pl', 'decide_group_phase', data, datafile=PREMIER_DATAFILE)
+    ranking = call_prolog('premier.pl', 'decide_group_phase', data, datafile=PREMIER_DATAFILE)
 
-    json_winners = jsonify_prolog(winners)
-    r = {'winners': json.loads(json_winners),
+    json_ranking = jsonify_prolog(ranking)
+    r = {'ranking': json.loads(json_ranking),
          'next_url': url_for('premier')}
     json_response = json.dumps(r)
     return json_response
 
+@app.route('/winner/<winner>', methods=['GET', 'POST'])
+def winner():
+    return {'winner': winner, 'next_url': url_for('winner', winner=winner)}
 
 def jsonify_prolog(string):
     '''Makes a prolog output string valid JSON'''
     # [[d,b],[h,f],[l,j],[p,n]] -> [["d","b"],["h","f"],["l","j"],["p","n"]]
-    json_string = re.sub(r'\b(\w+)\b', r'"\1"', string, flags=re.I)
+    json_string = re.sub(r'\b(\w+)\b', lambda m: '"{}"'.format(m.group(1).title()), string, flags=re.I)
     return json_string
 
 def prologify_json(string):
